Function skyway_webrtc_gateway_api::media::send_pli

pub async fn send_pli<'_, '_>(
    media_connection_id: &'_ MediaConnectionId, 
    params: &'_ SocketInfo<PhantomId>
) -> Result<(), Error>

Have WebRTC Gateway send a PLI(Picture Less Indication) packet

A PLI packets informs the encoder about the loss of an undefined amount of coded video data belonging to one or more pictures(RFC).

It's bindings for POST /media/connections/{media_connection_id}/pli

API

Examples

use skyway_webrtc_gateway_api::media::send_pli;
use skyway_webrtc_gateway_api::prelude::{MediaConnectionId, PhantomId, SerializableSocket, SocketInfo};

async fn example() {
    let media_connection_id = MediaConnectionId::new("mc-example");
    let socket = SocketInfo::<PhantomId>::new(None, "127.0.0.1:8000".parse().unwrap());
    let result = send_pli(&media_connection_id, &socket).await;
}