Module skyway_webrtc_gateway_api::peer

/peers api bindings

Structs

CreatePeerQuery
Query for POST /peers.

CreatedResponse
Response from POST /peers

PeerCallEvent
Shows that the Peer Object receives a request to establish MediaConnection with neighbour.

PeerCloseEvent
Response from GET /peers/{peer_id}/events

PeerConnectionEvent
Shows that the Peer Object receives a request to establish DataConnection with neighbour.

PeerErrorEvent
Response from GET /peers/{peer_id}/events

PeerInfo
Pair of PeerId and Token.

PeerOpenEvent
Response from GET /peers/{peer_id}/events

PeerStatusMessage
Response from GET /peer/{peer_id}/status

Enums

PeerEventEnum
Response from GET /peers/{peer_id}/events

Functions

create
Request to create Peer.

delete
Release PeerObject

listen_events
Listen events of a Peer Object.

status
Get status of PeerObject